F3D-25: close the Move transform-form integer domain
DesignMoveForm has the two source forms 1 and 5. Replace the raw native
integer and the parser's separate membership check with typed conversion.
Serialize each form as its original integer and reject other wire values
with an error naming form. No post-hoc Move-form validator existed.

The format defines the two codes but gives neither a distinct semantic
name, so the variants retain their source-code names. Parser tests keep
their expected source codes; exact integer round-trip and rejection tests
cover the native boundary. Scoped test compilation passes with and without
schema.

Hook-Bypass-Reason: User requested scoped compile checks between items and one final full gate.

+++ crates/cadmpeg-codec-f3d/src/design/decode/scopes/tests/legacy_frames.rs
@@ -268,7 +268,7 @@ fn legacy_move_transform_classes_use_the_shared_253_byte_envelope() {
 
         assert_eq!(decoded.transform, transform);
         assert_eq!(decoded.transform_record_index, record_index);
-        assert_eq!(decoded.form, form);
+        assert_eq!(u32::from(decoded.form), form);
         assert_eq!(decoded.form_offset, (frame_at + 43) as u64);
         assert_eq!(decoded.transform_offset, (frame_at + 48) as u64);
